chore: mark test as integration test
